ArchMd/requirements.txt
amqp==2.4.2
appnope==0.1.0
asn1crypto==0.24.0
autopep8==1.4.4
backcall==0.1.0
backports.csv==1.0.6
bcrypt==3.1.6
beautifulsoup4==4.7.1
billiard==3.6.0.0
bs4==0.0.1
celery==4.3.0
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
cryptography==2.6.1
decorator==4.4.0
defusedxml==0.5.0
diff-match-patch==20181111
Django==1.11.9
django-autocomplete-light==3.3.4
django-ckeditor==5.6.1
django-crispy-forms==1.7.2
django-debug-toolbar==1.11
django-formtools==2.1
django-haystack==2.8.1
django-import-export==1.2.0
django-js-asset==1.2.2
django-pure-pagination==0.3.0
django-redis==4.10.0
django-redis-sessions==0.6.1
django-reversion==3.0.3
django-silk==3.0.1
django-simple-captcha==0.5.5
django-tinymce==2.8.0
djangorestframework==3.9.2
djangorestframework-stubs==0.4.2
djdt-flamegraph==0.2.12
et-xmlfile==1.0.1
fabric==2.4.0
fdfs-client-py==1.2.6
future==0.17.1
gprof2dot==2016.10.13
httplib2==0.9.2
idna==2.8
invoke==1.2.0
ipython==7.4.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jdcal==1.4
jedi==0.13.3
jieba==0.39
Jinja2==2.10.1
kombu==4.5.0
MarkupSafe==1.1.1
mistune==0.8.4
mutagen==1.42.0
odfpy==1.4.0
openpyxl==2.6.1
paramiko==2.4.2
parso==0.4.0
pbr==5.1.3
pexpect==4.7.0
pickleshare==0.7.5
Pillow==5.4.1
prompt-toolkit==2.0.9
ptyprocess==0.6.0
pyasn1==0.4.5
pycodestyle==2.5.0
pycparser==2.19
pycryptodomex==3.7.2
Pygments==2.3.1
Pympler==0.7
PyMySQL==0.9.3
PyNaCl==1.3.0
python-alipay-sdk==1.10.0
python-dateutil==2.8.0
pytz==2018.9
PyYAML==3.13
redis==3.2.1
six==1.12.0
soupsieve==1.8
sqlparse==0.2.4
stevedore==1.30.1
tablib==0.13.0
traitlets==4.3.2
urllib3==1.24.1
uWSGI==2.0.18
vine==1.3.0
virtualenv==16.4.3
virtualenv-clone==0.5.1
virtualenvwrapper==4.8.4
wcwidth==0.1.7
Whoosh==2.7.4
xlrd==1.2.0
xlwt==1.3.0
